ai-persona/pyproject.toml
[project]
name = "ai-persona"
version = "0.1.0"
description = "AI Persona companion for IRL Twitch streamers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "twitchio>=3.1.0",
    "openai>=1.12.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tiktoken>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/persona", "config"]

[project.scripts]
ai-persona = "src.persona.main:run"
